Extract path from solution
This adds a new `generate-path` function that we can use to extract
the solution path from the `came-from` map.

I also renamed `visited` to `came-from` since that name seemed to
better reflect what it was doing. I also moved the `let` for `kids` down
below the early exit checks because it might be expensive to generate
the child nodes in some circumstances, and we wouldn't want to do that
if it wasn't necessary.